Finishing Pre school Education

+++ database/migrations/2017_12_20_083532_create_supplementary_food_distributions_table.php
@@ -18,7 +18,7 @@ class CreateSupplementaryFoodDistributionsTable extends Migration
             $table->integer('family_id');
             $table->integer('member_id');
             $table->smallInteger('target_type_id');
-            $table->tinyInteger('age');
+            $table->float('age');
             $table->integer('anganwadi_centre_id');
             $table->boolean('anganwadi_resident');
             $table->date('ration_given_date');
